feat, refactor, fix : 통합검색페이지(더보기) 기능 추가 및 Mapper.xml 리팩터링, FileService 수정 등

+++ src/main/java/org/zerock/wego/domain/common/FavoriteDTO.java
@@ -7,9 +7,9 @@ import lombok.Data;
 @Data
 public class FavoriteDTO {
 	
-	private long userId;
+	private Integer userId;
 	private String targetGb;
-	private long targetCd;
+	private Integer targetCd;
 	private Timestamp modifiedDt;
 	private Character status;
